[seg] clean up the code
