+++ src/storage/seg/segevict.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include "constant.h"
-//#include "seg.h"
 
 #include <time/time.h>
 #include <cc_mm.h>
@@ -27,12 +26,13 @@ typedef enum evict_rstatus {
 
 
 struct seg_evict_info {
-    evict_policy_e policy;
-    proc_time_i last_update_time;
-    int32_t nseg;
-    int32_t *ranked_seg_id;  /* the least valuable to the most valuable */
-    int32_t idx_rseg;     /* curr index in ranked seg id array */
-    pthread_mutex_t mtx;
+    evict_policy_e      policy;
+    proc_time_i         last_update_time;
+    int32_t             nseg;
+    int32_t             *ranked_seg_id;  /* ranked seg ids from the least
+                                          * valuable to the most valuable */
+    int32_t             idx_rseg;        /* curr index in ranked seg id array */
+    pthread_mutex_t     mtx;
 };
 
 
@@ -44,7 +44,8 @@ evict_rstatus_e
 least_valuable_seg(int32_t *seg_id);
 
 
-/* this must be setup update seg_setup has finished */
+/* this must be called after seg_setup has finished */
 void segevict_setup(evict_policy_e ev_policy, int32_t nseg);
 
 void segevict_teardown(void);
+
