+++ src/storage/seg/ttlbucket.h
@@ -12,93 +12,103 @@
 #define unlikely(x)    __builtin_expect(!!(x), 0)
 
 
-/*
- * segment based linked list memory allocator prioritizes TTL expiration over
- * eviction when cache space needs to be reclaimed.
+/**
+ * TTL indexed segment linked list, each segment (after allocation)
+ * is linked in one list, either a TTL bucket, or free pool
  *
- * We use an array of TTL buckets, each bucket stores the head of a segment list,
- * when memory needs to be reclaimed,
- * TODO(jason): need to think how to implement smart eviction
+ * This allows us to perform active TTL expiration - remove all expired
+ * objects without search the heap.
  *
- * slabclass[]:
+ * ** TTL buckets **
+ * We use an array of TTL buckets, each bucket stores the head and the tail of
+ * the segment list, insert only happens at the tail segment, when tail segment
+ * is full, we allocate a new seg and linked it into the list
  *
+ * because it is inefficient to have a list per ttl, so we bucket ttl into
+ * TTL buckets to reduce the number of ttl lists,
+ * specifically, we have the following TTL buckets (defined in constant.h)
+ *     1s  -    2047s (34m) : 256 buckets each 8s (except last bucket)
+ *   2048s -   32767s (9.1h): 256 buckets each 128s (first 16 bucket not used)
+ *  32768s -  524287s (6.1d): 256 buckets each 2048s (first 16 bucket not used)
+ * 524288s - 8388607s (97d) : 256 buckets each 32768s (first 16 bucket not used)
  *
- *
- *
- *
- *                               +------------------+             +------------------+
- *                               |                  |             |                  |
- *                               |                  |             |                  |
- *                               |     seg data     |             |     seg data     |
- *                               |                  |             |                  |
- *                               |                  |             |                  |
- *                               |                  |             |                  |
- * +----------------+            |                  |             |                  |
- * |                |            |                  |             |                  |
- * |  TTL bucket 1  |            +--------^---------+             +--------^---------+
- * |                |                     |                                |
- * |                |                     |                                |
- * +----------------+          +----------+----------+          +----------+----------+
- * |                |          |                     |          |                     |
- * |  TTL bucket 2  +---------->     struct seg      +---------->     struct seg      +--------->
- * |                |          |                     |          |                     |
- * |                |          +---------------------+          +---------------------+
- * +----------------+
- * |                |          +---------------------+          +---------------------+
- * |  TTL bucket 3  |          |                     |          |                     |
- * |                +---------->     struct seg      +---------->     struct seg      +--------->
- * |                |          |                     |          |                     |
- * +----------------+          +----------+----------+          +----------+----------+
- * |                |                     |                                |
- * |  TTL bucket 4  |                     |                                |
- * |                |                     |                                |
- * |                |            +--------v---------+             +--------v---------+
- * +----------------+            |                  |             |                  |
- *                               |                  |             |                  |
- *                               |    seg data      |             |    seg data      |
- *                               |                  |             |                  |
- *                               |                  |             |                  |
- *                               |                  |             |                  |
- *                               |                  |             |                  |
- *                               |                  |             |                  |
- *                               +------------------+             +------------------+
- *
- */
-
-
-/*
- * because it is inefficient to have a list per ttl, so we bucket ttl into ttl
- * buckets to reduce the number of ttl lists, specifically, we have the following
- * TTL buckets (defined in constant.h)
- *     1s   -     2047s (34m) :  256 buckets each 8s (except last bucket)
- *   2048s  -    32767s (9.1h):  256 buckets each 128s (first 16 bucket not used)
- *  32768s  -   524287s (6.1d):  256 buckets each 2048s (first 16 bucket not used)
- * 524288s  -  8388607s (97d) :  256 buckets each 32768s (first 16 bucket not used)
- *
- * a ttl 45 falls into bucket 5 (45 >> 3 = 5), a ttl 30000 falls into bucket 14,
+ * Example:
+ * ttl 45 falls into bucket 5 (45 >> 3 = 5),
+ * ttl 30000 falls into bucket 14
  * in total 1024 buckets are allocated at start up, and
  * (16+40+8) B * 1024 = 32 KiB DRAM will be consumed
  *
- * NOTE: there are 48 buckets wasted,  it can be used if needed,
- * but the the code complexity will skyrocket
+ * currently there are 48 buckets wasted,  it can be used if needed
+ *
+ *
+ * ** segment headers **
+ * we use an array of segment headers, the size of array is the same as the
+ * number of segments, each header corresponds to one fixed segment,
+ * taking the segment header out of segment has two benefits:
+ * 1. because seg header is small, it can be cached in L3 more easily (not confirmed)
+ * 2. we update metadata of each segment frequently, separating seg headers
+ *      allow us to use slower device for segments (PMem, SSD) with
+ *      better performance and low write amplification
+ *
+ *
+ *                                       segment header array
+ *                                    ┌────────────────────────┐
+ *                                    │                        │
+ *                                ┌──▶│    segment header 1    ├──next┐
+ *                                │   │                        │      │
+ *                                │   ├────────────────────────┤      │
+ *                                │   │                        │      │
+ *                                │   │    segment header 2    ┣ ━ ━ ━│━ ━
+ *  TTL bucket array              │   │                        │      │   ┃
+ * ┌────────────────┐  first seg  │   ├────────────────────────┤      │
+ * │                ├─────────────┘   │                        │◀─────┘   ┃
+ * │  TTL bucket 1  │                 │    segment header 3    │──next┐
+ * │                ├─────────────┐   │                        │      │   ┃
+ * ├────────────────┤  last seg   │   ├────────────────────────┤      │
+ * │                │             └──▶│                        │      │   ┃
+ * │  TTL bucket 2  │                 │          ...           │◀─────┘
+ * │                │                 │                        │          ┃
+ * ├────────────────┤                 ├────────────────────────┤
+ * │                │                 │                        │          ┃
+ * │      ...       │             ┌──▶│          ...           │
+ * │                │             │   │                        │          ┃
+ * ├────────────────┤             │   ├────────────────────────┤
+ * │                │             │   │                        │◀ ━ ━ ━ ━ ┛
+ * │      ...       │             │   │          ...           │━ ━ ━ ━ ━ ┓
+ * │                │             │   │                        │
+ * ├────────────────┤  first seg  │   ├────────────────────────┤          ┃
+ * │                ├─────────────┘   │                        │
+ * │ TTL bucket 1022│                 │          ...           │          ┃
+ * │                ├─────────────┐   │                        │
+ * ├────────────────┤  last seg   │   ├────────────────────────┤          ┃
+ * │                │             │   │                        │
+ * │ TTL bucket 1023│             │   │          ...           │◀ ━ ━ ━ ━ ┛
+ * │                │             │   │                        │
+ * └────────────────┘             │   ├────────────────────────┤
+ *                                │   │                        │
+ *                                └──▶│   segment header N-2   │
+ *                                    │                        │
+ *                                    ├────────────────────────┤
+ *                                    │                        │
+ *                                    │   segment header N-1   │
+ *                                    │                        │
+ *                                    ├────────────────────────┤
+ *                                    │                        │
+ *                                    │    segment header N    │
+ *                                    │                        │
+ *                                    └────────────────────────┘
+ *
+ *
  */
 
 
-//TAILQ_HEAD(seg_id_tqh, uint32_t);
-//TAILQ_ENTRY(uint32_t) seg_id_tqe;
-
-/* we can use a timing wheel here or in seg */
 struct ttl_bucket {
     int32_t first_seg_id;
     int32_t last_seg_id;
-
-    pthread_mutex_t mtx;
-
     delta_time_i    ttl;           /* the min ttl of this bucket */
     uint32_t n_seg;
 };
 
-extern struct ttl_bucket ttl_buckets[MAX_TTL_BUCKET];
 
 /**
  * give a TTL, find the index of TTL bucket in ttl_array
@@ -134,11 +144,6 @@ find_ttl_bucket_idx(delta_time_i ttl)
     return bucket_idx;
 }
 
-/* return the (min) TTL of the bucket */
-static inline delta_time_i
-bucket_idx_to_ttl(uint16_t bucket_idx){
-    return ttl_buckets[bucket_idx].ttl;
-}
 
 void
 ttl_bucket_setup(void);
@@ -146,17 +151,12 @@ ttl_bucket_setup(void);
 void
 ttl_bucket_teardown(void);
 
-/*
+/**
  * Reserve an item from the active segment of the ttl bucket.
  * If the active seg of current ttl bucket does not have enough space,
- * we first try to get an empty segment,
- * if there is no free seg,
- *      we check whether there are expired segment,
- *      if yes, remove the seg (refcount must be 0),
- *      if no, we evict one seg,
- * then we link the seg into ttl_bucket, make it the current active seg.
+ * we will get an empty segment
+ * then link the seg into ttl_bucket, make it the current active seg.
  *
  */
 struct item *
 ttl_bucket_reserve_item(int32_t ttl_bucket_idx, size_t sz, int32_t *seg_id);
-
